unfinished 1.21.8 port

+++ src/main/kotlin/com/github/synnerz/talium/utils/ScaledResolution.kt
@@ -5,7 +5,7 @@ import net.minecraft.client.MinecraftClient
 class ScaledResolution(val mc: MinecraftClient) {
     val scaledWidth = mc.window.scaledWidth
     val scaledHeight = mc.window.scaledHeight
-    val scaleFactor = mc.window.scaleFactor.toInt()
+    val scaleFactor = mc.window.scaleFactor
     val scaledWidth_double get() = scaledWidth.toDouble()
     val scaledHeight_double get() = scaledHeight.toDouble()
 }
